requirements.txt
boto3==1.7.62
dj-database-url==0.5.0
Django==2.1.5
django-forms-bootstrap==3.1.0
django-materializecss-form==1.1.11
django-storages==1.6.6
djangorestframework==3.9.1
djangorestframework-stubs==0.2.0
gunicorn==19.9.0
numpy==1.16.2
pandas==0.24.1
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2018.9
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
xgboost==0.82
